requirements.txt
appdirs==1.4.3
bitshares==0.1.22
certifi==2018.8.24
chardet==3.0.4
ecdsa==0.13
Events==0.3
future==0.16.0
graphenelib==0.6.6
idna==2.7
pycryptodome==3.6.6
pylibscrypt==1.7.1
python-telegram-bot==8.1.1
requests==2.19.1
scrypt==0.8.6
six==1.11.0
urllib3==1.23
websocket-client==0.48.0
websockets==6.0
